services/enriquecimento/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
python-dotenv
prometheus-fastapi-instrumentator
